+++ .storybook/main.ts
@@ -141,10 +141,9 @@ async function ignoreServerFunctionModules(
   webpackConfig.plugins.push(
     new IgnorePlugin({
       checkResource(resource) {
-        return [
-          './createSnapshot',
-          '@/app/backups/new/createNewBackup',
-        ].includes(resource)
+        return ['./createSnapshot', './backups/new/createNewBackup'].includes(
+          resource
+        )
       },
     })
   )
