+++ src/app/LoggedIn.tsx
@@ -2,17 +2,33 @@
 
 import { Account, Client, useAuthenticator } from '@storacha/ui-react'
 import { styled } from 'next-yak'
-import { useEffect, useState } from 'react'
+import { ReactNode, useEffect, useState } from 'react'
+import { useFormStatus } from 'react-dom'
+import { toast } from 'sonner'
 
 import { BackupDetail } from '@/components/BackupScreen/BackupDetail'
 import { Center, Stack, Text } from '@/components/ui'
+import { CreateButton } from '@/components/ui/CreateButton'
 import { atproto } from '@/lib/capabilities'
 import { SERVER_DID } from '@/lib/constants'
+import { delegate } from '@/lib/delegate'
+import { uploadCAR } from '@/lib/storacha'
+import { SpaceDid } from '@/types'
 
 import { BackupScreen } from '../components/BackupScreen'
 import { useSWR } from '../lib/swr'
 
 import { Sidebar } from './Sidebar'
+
+let createNewBackup: typeof import('./backups/new/createNewBackup').action
+
+if (process.env.STORYBOOK) {
+  createNewBackup = () => {
+    throw new Error('Server Functions are not available in Storybook')
+  }
+} else {
+  createNewBackup = (await import('./backups/new/createNewBackup')).action
+}
 
 const Outside = styled(Stack)`
   min-height: 100vh;
@@ -39,6 +55,65 @@ async function createSession(client: Client, account: Account) {
     method: 'POST',
     body: archivedDelegation,
   })
+}
+
+function NewBackupForm({ children }: { children: ReactNode }) {
+  const [{ client }] = useAuthenticator()
+
+  async function generateDelegationAndCreateNewBackup(formData: FormData) {
+    try {
+      const space = formData.get('storacha_space') as SpaceDid | undefined
+
+      if (!space) {
+        console.error('space id not defined, cannot create delegation.')
+        toast.error('Space ID not defined, cannot create delegation.')
+        return
+      }
+
+      if (!client) {
+        console.error('client not defined, cannot create delegation')
+        toast.error('Client not defined, cannot create delegation.')
+        return
+      }
+
+      await client.setCurrentSpace(space)
+      // upload the delegation to Storacha so we can use it later
+
+      // Create a delegation valid for a year of backups
+      const delegationDuration = 1000 * 60 * 60 * 24 * 365
+      const delegationCid = await uploadCAR(
+        client,
+        new Blob([
+          await delegate(client, space, { duration: delegationDuration }),
+        ])
+      )
+      formData.append('delegation_cid', delegationCid.toString())
+
+      const result = await createNewBackup(formData)
+      return result
+    } catch (error) {
+      toast.error(
+        error instanceof Error
+          ? error.message === 'NEXT_REDIRECT'
+            ? 'Backup created successfully! Redirecting...'
+            : error.message
+          : 'Unknown error'
+      )
+      console.error('Backup creation error:', error)
+    }
+  }
+
+  return <form action={generateDelegationAndCreateNewBackup}>{children}</form>
+}
+
+const CreateBackupButton = () => {
+  const { pending } = useFormStatus()
+
+  return (
+    <CreateButton $isLoading={pending} type="submit">
+      Create Backup
+    </CreateButton>
+  )
 }
 
 export function LoggedIn() {
@@ -76,7 +151,10 @@ export function LoggedIn() {
           </Center>
         }
       >
-        <BackupDetail account={account} />
+        <NewBackupForm>
+          <BackupDetail account={account} />
+          <CreateBackupButton />
+        </NewBackupForm>
       </BackupScreen>
     </Outside>
   )
