+++ src/components/BackupScreen/BackupDetail.tsx
@@ -1,34 +1,16 @@
 'use client'
 
-import { Account, useAuthenticator } from '@storacha/ui-react'
+import { Account } from '@storacha/ui-react'
 import { styled } from 'next-yak'
 import { ReactNode, useEffect, useState } from 'react'
-import { useFormStatus } from 'react-dom'
-import { toast } from 'sonner'
 
 import { BlueskyAccountSelect } from '@/components/BackupScreen/BlueskyAccountSelect'
 import { StorachaSpaceSelect } from '@/components/BackupScreen/StorachaSpaceSelect'
-import { CreateButton } from '@/components/ui/CreateButton'
-import { delegate } from '@/lib/delegate'
-import { uploadCAR } from '@/lib/storacha'
+import { Heading, Stack, Text } from '@/components/ui'
 import { shortenDID } from '@/lib/ui'
-import { Backup, SpaceDid } from '@/types'
-
-import { Container, Heading, Stack, Text } from '../ui'
+import { Backup } from '@/types'
 
 import { DataBox } from './Data'
-
-let createNewBackup: typeof import('@/app/backups/new/createNewBackup').action
-
-if (process.env.STORYBOOK) {
-  createNewBackup = () => {
-    throw new Error('Server Functions are not available in Storybook')
-  }
-} else {
-  // Addressed further on in https://github.com/storacha/bluesky-backup-webapp-server/pull/101
-  // eslint-disable-next-line import/no-restricted-paths
-  createNewBackup = (await import('@/app/backups/new/createNewBackup')).action
-}
 
 interface BackupProps {
   account?: Account
@@ -97,75 +79,6 @@ const DATA_BOXES: DataConfig[] = [
   // },
 ]
 
-const CreateBackupButton = () => {
-  const { pending } = useFormStatus()
-
-  return (
-    <CreateButton $isLoading={pending} type="submit">
-      create backup
-    </CreateButton>
-  )
-}
-
-function NewBackupForm({ children }: { children: ReactNode }) {
-  const [{ client }] = useAuthenticator()
-
-  async function generateDelegationAndCreateNewBackup(formData: FormData) {
-    try {
-      const space = formData.get('storacha_space') as SpaceDid | undefined
-
-      if (!space) {
-        console.error('space id not defined, cannot create delegation.')
-        toast.error('Space ID not defined, cannot create delegation.')
-        return
-      }
-
-      if (!client) {
-        console.error('client not defined, cannot create delegation')
-        toast.error('Client not defined, cannot create delegation.')
-        return
-      }
-
-      await client.setCurrentSpace(space)
-      // upload the delegation to Storacha so we can use it later
-
-      // Create a delegation valid for a year of backups
-      const delegationDuration = 1000 * 60 * 60 * 24 * 365
-      const delegationCid = await uploadCAR(
-        client,
-        new Blob([
-          await delegate(client, space, { duration: delegationDuration }),
-        ])
-      )
-      formData.append('delegation_cid', delegationCid.toString())
-
-      const result = await createNewBackup(formData)
-      return result
-    } catch (error) {
-      toast.error(
-        error instanceof Error
-          ? error.message === 'NEXT_REDIRECT'
-            ? 'Backup created successfully! Redirecting...'
-            : error.message
-          : 'Unknown error'
-      )
-      console.error('Backup creation error:', error)
-    }
-  }
-
-  return <form action={generateDelegationAndCreateNewBackup}>{children}</form>
-}
-
-function MaybeForm({
-  children,
-  backup,
-}: {
-  children: ReactNode
-  backup?: Backup
-}) {
-  return backup ? children : <NewBackupForm>{children}</NewBackupForm>
-}
-
 const Section = ({
   title,
   children,
@@ -205,58 +118,51 @@ export const BackupDetail = ({ account, backup }: BackupProps) => {
 
   return (
     <>
-      <MaybeForm backup={backup}>
-        {account && (
-          <input type="hidden" name="account" value={account.did()} />
+      {account && <input type="hidden" name="account" value={account.did()} />}
+      <Stack $gap="2rem">
+        {backup ? (
+          <Heading>{backup.name}</Heading>
+        ) : (
+          <Heading>New Backup</Heading>
         )}
-        <Container>
-          <Stack $gap="2rem">
-            {backup ? (
-              <Heading>{backup.name}</Heading>
-            ) : (
-              <Heading>New Backup</Heading>
-            )}
-            <Section title="Accounts">
-              <AccountsContainer $direction="row">
-                <Wrapper>
-                  <BlueskyAccountSelect
-                    name="atproto_account"
-                    {...(backup && {
-                      disabled: true,
-                      value: backup.atprotoAccount,
-                    })}
-                  />
-                </Wrapper>
-                <Wrapper>
-                  <StorachaSpaceSelect
-                    name="storacha_space"
-                    {...(backup && {
-                      disabled: true,
-                      value: shortenDID(backup.storachaSpace),
-                    })}
-                  />
-                </Wrapper>{' '}
-              </AccountsContainer>
-            </Section>
+        <Section title="Accounts">
+          <AccountsContainer $direction="row">
+            <Wrapper>
+              <BlueskyAccountSelect
+                name="atproto_account"
+                {...(backup && {
+                  disabled: true,
+                  value: backup.atprotoAccount,
+                })}
+              />
+            </Wrapper>
+            <Wrapper>
+              <StorachaSpaceSelect
+                name="storacha_space"
+                {...(backup && {
+                  disabled: true,
+                  value: shortenDID(backup.storachaSpace),
+                })}
+              />
+            </Wrapper>{' '}
+          </AccountsContainer>
+        </Section>
 
-            <Section title="Data">
-              <Stack $direction="row" $gap="1.25rem" $wrap="wrap">
-                {DATA_BOXES.map((box) => (
-                  <DataBox
-                    key={box.key}
-                    name={box.name}
-                    title={box.title}
-                    description={box.description}
-                    value={data[box.key] || false}
-                    onToggle={() => !backup && toggle(box.key)}
-                  />
-                ))}
-              </Stack>
-            </Section>
-            {!backup && <CreateBackupButton />}
+        <Section title="Data">
+          <Stack $direction="row" $gap="1.25rem" $wrap="wrap">
+            {DATA_BOXES.map((box) => (
+              <DataBox
+                key={box.key}
+                name={box.name}
+                title={box.title}
+                description={box.description}
+                value={data[box.key] || false}
+                onToggle={() => !backup && toggle(box.key)}
+              />
+            ))}
           </Stack>
-        </Container>
-      </MaybeForm>
+        </Section>
+      </Stack>
     </>
   )
 }
